Refactor: Extract  new mod error
